fix unloaded prefetched content
